main: convert last_error_lineno global to type uint32_t
As this is the expected type for lineno everywhere else in the engine

+++ main/main.c
@@ -1344,7 +1344,7 @@ static ZEND_COLD void php_error_cb(int orig_type, zend_string *error_filename, c
 		 * be NULL if PG(last_error_message) is not NULL */
 		if (!zend_string_equals(PG(last_error_message), message)
 			|| (!PG(ignore_repeated_source)
-				&& ((PG(last_error_lineno) != (int)error_lineno)
+				&& ((PG(last_error_lineno) != error_lineno)
 					|| !zend_string_equals(PG(last_error_file), error_filename)))) {
 			display = 1;
 		} else {
